gemini-streamlit-cloudrun/embedding_utils/requirements.txt
transformers==4.40.1
torch==2.2.2
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.2
